disallow passing arrays of tuples as arguments to `%`
Nim converts {"key":"val"} to [("key", "val")] at compile time. So {} is a tuple inside an array actually.
Using the % operator on this will pass it as an argument to the respective function that takes an openarray as argument and convert the tuple inside this array:
%{"key":"val"} becomes [{"Field0":"key","Field1":"val"}].
As to why this is a problem, I do not know, but two tests failed because of that. 
Before, this wasn't a problem, because one of the `%` procedures that was supposed to accept tuples and convert them to JsonNodes only accepted objects. Now that we also want to convert tuples to a JSON object as well via `%`, we have to make sure that {} is not handled as a tuple inside an array also.